feat: config layout and login page
